refactor doozer images:* commands
